% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R, R/hello.R
\name{fibonacci}
\alias{fibonacci}
\title{Compute fibonacci sequence}
\usage{
fibonacci(n)

fibonacci(n)
}
\description{
when the first and second sections are 1,
from the third section is the sum of the first two sections.

when the first and second sections are 1,
from the third section is the sum of the first two sections.
}
\examples{


}
